+++ src/models/agents.ts
@@ -2,6 +2,6 @@ export interface GridItem {
     data: { id: number; x: number; y: number }
     style: {
         image?: string,
-        backgroundColor?: string
+        color?: string
     }
 }