+++ ReleaseNotes.md
@@ -12,6 +12,12 @@
 - Echarts server-side configurable items
 - Configure connections to different types of databases by connection string, and get data by sqlalchemy.
 
+## Version 0.5.0
+- Feature: Fixed network and grid visualization;
+- Feature: Dynamic actions from Melodie visualizer to webpage in order to implement customized actions, eg. export table to `.xlsx` file from database.
+- Enhancement: Added border to graph and visualization components.
+
+
 ## Version 0.4.0
 - Feature: Complexed tree parameters
 - Feature: Full paged visualizer for showing on the big screen